+++ templates/admin-panel/template/client/src/services/store/api.js
@@ -1,0 +1,70 @@
+import { inject } from 'sham-ui-macro/babel.macro';
+import axios from 'axios';
+
+export class API {
+    @inject session;
+
+    constructor( { baseURL, onUnauthorized } ) {
+        this.axios = axios.create( {
+            baseURL,
+            withCredentials: true
+        } );
+        this.axios.interceptors.request.use(
+            ::this._requestInterceptor
+        );
+        this.axios.interceptors.response.use(
+            ::this._responseInterceptor,
+            ::this._responseFailInterceptor
+        );
+
+        this._onUnauthorized = onUnauthorized;
+    }
+
+    request( config ) {
+        return this.axios.request( config );
+    }
+
+    _requestInterceptor( request ) {
+        if ( this.token ) {
+            request.headers[ 'X-CSRF-Token' ] = this.token;
+        }
+        return request;
+    }
+
+    _responseInterceptor( response ) {
+        const token = response.headers[ 'x-csrf-token' ];
+        if ( token ) {
+
+            // Save token from response
+            this.token = token;
+        }
+
+        // Return only data for success response
+        return response.data;
+    }
+
+    _responseFailInterceptor( error ) {
+        const { response, config } = error;
+        if (
+            this.session.data.isAuthenticated &&
+            response &&
+            401 === response.status
+        ) {
+
+            // Handle 401 HTTP status code for non authorized user
+            this._onUnauthorized( {
+                ...config,
+                url: config.url.replace( config.baseURL, '/' )
+            } );
+        }
+        return Promise.reject(
+            this.constructor.extractErrors( error )
+        );
+    }
+
+    static extractErrors( error ) {
+        return error && error.response && error.response.data ?
+            error.response.data :
+            {};
+    }
+}