+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## master
 * Update `sham-ui` dependencies
+* Refactoring store service in admin-panel [#84](https://github.com/sham-ui/sham-ui-cli/issues/84) 
+* Fix a tests
 
 ## 4.3.0
 * Add template for `styled-component` [#3](https://github.com/sham-ui/sham-ui-cli/issues/3)